organize, fix passing of apify token
