+++ src/mcp/server.ts
